feat(traces): keep closed trades' order traces in a local archive and resolve them through one path
The core answers `request_traces(ReportUID)` from an archive it keeps for a
bounded time, and MoonProto keeps no cache: the protocol asks the terminal to
store what it received by `ReportUID` and to ask only for what it does not
have. Until now the answers lived in `CoreData::report_traces` for the
session, the trade window asked the core on every open, and a trade the core
had aged out was gone for good.

- `data/order_traces.sqlite` (`db::order_traces`): one writer thread, tables
  `trace_answers` (per row: check time, line count) and `trace_lines` (points
  as f64 ms / f64 price, the archive is not downcast); an empty answer is
  filed with its date and re-asked only after 7 days; a failed request is
  never filed. The file joins the startup core-uid floor.
- the feed files every `TraceReady` in the archive, queues an ask when a
  report row closes (`RowUpsert` with `CloseDate != 0`, field indices cached
  per schema revision, the writer decides whether the row is already known)
  and, after a core's `SyncComplete`, backfills the closed rows of the last
  30 days without a trace, newest first, up to 400 per start
- every `request_traces` of a connection goes through one paced queue
  (`live::trace_backfill`): at most 8 unanswered, one send per 200 ms, a
  window's ask ahead of the backfill; three failures in a row abandon the
  queue (an old core never answers)
- one trace resolver on `Backend` (`backend::traces`, wake channel
  `TracesRevision`): archive read off-thread first, the core for misses up to
  a caller's cap, answers adopted from the session store on the feed drain
  edge (`DrainStats::report_traces`, which no longer repaints the backend's
  observers) and on the slow tick for a removed core
- the trade window consumes the resolver instead of polling the store itself;
  `ArchivedOrderTrace` prices become f64, narrowed to f32 only when a
  `LineTrace` is built

Verified on 26 live cores: 4620 backfill asks answered in 82 s at 5/s per
core with zero failures; a restart listed 0 rows for the 21 cores that were
fully filed and continued past the cap on the other 5.

Refs #621

+++ crates/moon-core/src/session/order_lines/archived/tests.rs
@@ -1,6 +1,6 @@
 use super::*;
 
-fn trace(own: bool, kind: ArchivedLineKind, points: &[(f64, f32)]) -> ArchivedOrderTrace {
+fn trace(own: bool, kind: ArchivedLineKind, points: &[(f64, f64)]) -> ArchivedOrderTrace {
     ArchivedOrderTrace {
         own,
         kind,
